+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.0] - 2019-05-07
+### Added
+- Support for multiple languages in translations
+- Language selector component
+- Simple grid primitive with optional forced padding on children nodes
+
+### Changed
+- Structure of translations object
+- Documentation of translation support
+
+### Fixed
+- `getNodeTitles` function in selectors now build using the id, not the property name as was the case before
+
 ## [4.2.0] - 2019-02-20
 
 ### Added
@@ -122,7 +135,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.2] -
 
-[unreleased]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v4.2.0...HEAD
+[unreleased]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v5.0.0...HEAD
+[5.0.0]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v4.2.0...v5.0.0
 [4.2.0]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v4.1.3...v4.2.0
 [4.1.3]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v4.1.2...v4.1.3
 [4.1.2]: https://github.com/DirektoratetForByggkvalitet/losen/compare/v4.1.1...v4.1.2